+++ cookbooks/spacewalk-nativex/attributes/default.rb
@@ -1,0 +1,4 @@
+# Clones/Merges Dev channel packages into Prod channel
+default["spacewalk-nativex"]["update_prod"] = true
+# Backs up Prod channel before cloning/merging Dev packages
+default["spacewalk-nativex"]["backup_prod"] = true