+++ cookbooks/spacewalk-nativex/metadata.rb
@@ -1,7 +1,10 @@
-name             'spacewalkclient-nativex'
+name             'spacewalk-nativex'
 maintainer       'NativeX'
 maintainer_email 'adrian.herrera@nativex.com'
 license          'All rights reserved'
-description      'Installs/Configures spacewalkclient-nativex'
+description      'Installs/Configures spacewalk-nativex'
 long_description IO.read(File.join(File.dirname(__FILE__), 'README.md'))
 version          '0.1.0'
+
+depends 'cron'
+#depends 'ad-nativex'